Port the installer to paramiko + Change installs to local HTTP server
