+++ robotpy_installer/__init__.py
@@ -1,0 +1,1 @@
+__version__ = "2019.2.0"
